Renamed RealmFileWrapper to SharedGroupManager and RealmInstance to RealmBase + other refactors for cleaner separation
