+++ realm/src/main/java/io/realm/internal/RealmBase.java
@@ -21,9 +21,11 @@ import android.os.Handler;
 import android.os.Looper;
 import android.os.Message;
 
-import java.io.*;
+import java.io.Closeable;
+import java.io.File;
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -45,12 +47,17 @@ import io.realm.internal.log.RealmLog;
  * class. This means that all classes that you want JavaDoc for should be implemented in the
  * subclass :(
  */
-public abstract class RealmInstance implements Closeable {
+public abstract class RealmBase implements Closeable {
 
     protected static final int REALM_CHANGED = 14930352; // Hopefully it won't clash with other message IDs.
     protected static final String INCORRECT_THREAD_CLOSE_MESSAGE = "Realm access from incorrect thread. Realm instance can only be closed on the thread it was created.";
     private static final String INCORRECT_THREAD_MESSAGE = "Realm access from incorrect thread. Realm objects can only be accessed on the thread they were created.";
     private static final String CLOSED_REALM_MESSAGE = "This Realm instance has already been closed, making it unusable.";
+    private static final String DIFFERENT_KEY_MESSAGE = "Wrong key used to decrypt Realm.";
+
+    // Map between all Realm file paths and all known configurations pointing to that file.
+    private static final Map<String, List<RealmConfiguration>> globalPathConfigurationCache =
+            new HashMap<String, List<RealmConfiguration>>();
 
     // Reference count for how many open Realm instances there currently are on this thread.
     protected static final ThreadLocal<Map<RealmConfiguration, Integer>> referenceCount =
@@ -69,7 +76,7 @@ public abstract class RealmInstance implements Closeable {
 
     protected long threadId;
     protected RealmConfiguration configuration;
-    protected RealmFileWrapper realmFile;
+    protected SharedGroupManager sharedGroup;
     protected boolean autoRefresh;
     protected Handler handler;
 
@@ -77,10 +84,10 @@ public abstract class RealmInstance implements Closeable {
         RealmLog.add(BuildConfig.DEBUG ? new DebugAndroidLogger() : new ReleaseAndroidLogger());
     }
 
-    public RealmInstance(RealmConfiguration configuration, boolean autoRefresh) {
+    public RealmBase(RealmConfiguration configuration, boolean autoRefresh) {
         this.threadId = Thread.currentThread().getId();
         this.configuration = configuration;
-        this.realmFile = new RealmFileWrapper(configuration);
+        this.sharedGroup = SharedGroupManager.getInstance(threadId, configuration);
         setAutoRefresh(autoRefresh);
     }
 
@@ -147,7 +154,7 @@ public abstract class RealmInstance implements Closeable {
 
     public void checkIfValid() {
         // Check if the Realm instance has been closed
-        if (realmFile != null && !realmFile.isOpen()) {
+        if (sharedGroup != null && !sharedGroup.isOpen()) {
             throw new IllegalStateException(CLOSED_REALM_MESSAGE);
         }
 
@@ -186,6 +193,7 @@ public abstract class RealmInstance implements Closeable {
      * Closes a single instance of a Realm.
      * @return {@code true} if it was the last instance close, {@code false} otherwise.
      */
+    // TODO Not threadsafe
     protected boolean closeInstance(RealmConfiguration configuration) {
         if (this.threadId != Thread.currentThread().getId()) {
             throw new IllegalStateException(INCORRECT_THREAD_CLOSE_MESSAGE);
@@ -201,9 +209,9 @@ public abstract class RealmInstance implements Closeable {
 
         if (references == 1) {
             wasLastInstance = true;
-            realmFile.close();
-            realmFile = null;
-            RealmFileWrapper.closeFile(configuration);
+            sharedGroup.close();
+            sharedGroup = null;
+            globalPathConfigurationCache.get(canonicalPath).remove(configuration);
         }
 
         int refCount = references - 1;
@@ -228,22 +236,22 @@ public abstract class RealmInstance implements Closeable {
             throw new IllegalArgumentException("The destination argument cannot be null");
         }
         checkIfValid();
-        realmFile.copyToFile(destination, key);
+        sharedGroup.copyToFile(destination, key);
     }
 
     protected void refresh() {
         checkIfValid();
-        realmFile.advanceRead();
+        sharedGroup.advanceRead();
     }
 
     protected void beginTransaction() {
         checkIfValid();
-        realmFile.promoteToWrite();
+        sharedGroup.promoteToWrite();
     }
 
     protected void commitTransaction() {
         checkIfValid();
-        realmFile.commitAndContinueAsRead();
+        sharedGroup.commitAndContinueAsRead();
 
         for (Map.Entry<Handler, String> handlerIntegerEntry : handlers.entrySet()) {
             Handler handler = handlerIntegerEntry.getKey();
@@ -268,7 +276,7 @@ public abstract class RealmInstance implements Closeable {
 
     protected void cancelTransaction() {
         checkIfValid();
-        realmFile.rollbackAndContinueAsRead();
+        sharedGroup.rollbackAndContinueAsRead();
     }
 
     protected String getPath() {
@@ -279,11 +287,67 @@ public abstract class RealmInstance implements Closeable {
         return configuration;
     }
 
+
+    /**
+     * Make sure that the new configuration doesn't clash with any existing configurations for the
+     * Realm.
+     *
+     * @throws IllegalArgumentException If the new configuration isn't valid.
+     */
+    protected static void validateAgainstExistingConfigurations(RealmConfiguration newConfiguration) {
+
+        // Ensure cache state
+        String realmPath = newConfiguration.getPath();
+        List<RealmConfiguration> pathConfigurationCache = globalPathConfigurationCache.get(realmPath);
+        if (pathConfigurationCache == null) {
+            pathConfigurationCache = new CopyOnWriteArrayList<RealmConfiguration>();
+            globalPathConfigurationCache.put(realmPath, pathConfigurationCache);
+        }
+
+        if (pathConfigurationCache.size() > 0) {
+
+            // For the current restrictions, it is enough to just check one of the existing configurations.
+            RealmConfiguration cachedConfiguration = pathConfigurationCache.get(0);
+
+            // Check that encryption keys aren't different
+            if (!Arrays.equals(cachedConfiguration.getEncryptionKey(), newConfiguration.getEncryptionKey())) {
+                throw new IllegalArgumentException(DIFFERENT_KEY_MESSAGE);
+            }
+
+            // Check schema versions are the same
+            if (cachedConfiguration.getSchemaVersion() != newConfiguration.getSchemaVersion()) {
+                throw new IllegalArgumentException(String.format("Configurations cannot have different schema versions " +
+                                "if used to open the same file. %d vs. %d", cachedConfiguration.getSchemaVersion(),
+                        newConfiguration.getSchemaVersion()));
+            }
+
+            // Check that schema is the same
+            RealmProxyMediator cachedSchema = cachedConfiguration.getSchemaMediator();
+            RealmProxyMediator schema = newConfiguration.getSchemaMediator();
+            if (!cachedSchema.equals(schema)) {
+                throw new IllegalArgumentException("Two configurations with different schemas are trying to open " +
+                        "the same Realm file. Their schema must be the same: " + newConfiguration.getPath());
+            }
+
+            // Check if the durability is the same
+            SharedGroup.Durability cachedDurability = cachedConfiguration.getDurability();
+            SharedGroup.Durability newDurability = newConfiguration.getDurability();
+            if (!cachedDurability.equals(newDurability)) {
+                throw new IllegalArgumentException("A Realm cannot be both in-memory and persisted. Two conflicting " +
+                        "configurations pointing to " + newConfiguration.getPath() + " are being used.");
+            }
+        }
+
+        // The new configuration doesn't violate existing configurations. Cache it.
+        pathConfigurationCache.add(newConfiguration);
+    }
+
+
     private class RealmCallback implements Handler.Callback {
         @Override
         public boolean handleMessage(Message message) {
             if (message.what == REALM_CHANGED) {
-                realmFile.advanceRead();
+                sharedGroup.advanceRead();
                 sendNotifications();
             }
             return true;
